+++ lib/event_db.dart
@@ -2,5 +2,6 @@ library event_db;
 
 export 'src/model.dart';
 export 'src/model_map.dart';
+export 'src/reorderable.dart';
 export 'src/repository.dart';
 export 'src/fake.dart';